requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Topological Data Analysis
giotto-tda>=0.6.0
persim>=0.3.1

# Configuration
pyyaml>=6.0

# Logging & Visualization
tensorboard>=2.13.0
tqdm>=4.65.0

# Scientific Computing
scikit-learn>=1.2.0
pandas>=2.0.0

# Testing
pytest>=7.3.0
